server/requirements.txt
anyio==4.2.0
beautifulsoup4==4.12.3
blinker==1.6.2
Brotli==1.0.9
cachelib==0.10.2
certifi==2022.9.24
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
deta==1.2.0
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Session==0.5.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
markdown-it-py==3.0.0
marketwatch==0.2.2
MarkupSafe==2.1.3
mdurl==0.1.2
packaging==23.2
Pygments==2.17.2
python-dotenv==1.0.1
requests==2.31.0
rich==13.7.0
sniffio==1.3.0
soupsieve==2.5
urllib3==2.0.5
Werkzeug==2.3.7
